Improve CfgBackEndFetchCmds.MaxCmds handling
Summary:
- Set ap.configd to fetch more than one command
- Clarify the meaning of MaxCmds == 0 (it means: default to 1)
- MaxCmds should be a uint32 throughout the stack

Test Plan:
- Manual testing of cl.configd, ap.configd.
- Sent commands up and down

Reviewers: nils, duvall

Reviewed By: nils, duvall

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D435

+++ golang/src/bg/cl.configd/configd.go
@@ -43,7 +43,7 @@ type configStore interface {
 
 type cmdQueue interface {
 	submit(context.Context, *perAPState, *cfgmsg.ConfigQuery) (int64, error)
-	fetch(context.Context, *perAPState, int64, int64) ([]*cfgmsg.ConfigQuery, error)
+	fetch(context.Context, *perAPState, int64, uint32) ([]*cfgmsg.ConfigQuery, error)
 	status(context.Context, *perAPState, int64) (*cfgmsg.ConfigResponse, error)
 	cancel(context.Context, *perAPState, int64) (*cfgmsg.ConfigResponse, error)
 	complete(context.Context, *perAPState, *cfgmsg.ConfigResponse) error
